builds are created
Since builds cannot be re-executed, a Source will create and manage N Builds.

informerfix is to register the build informers with the knative.dev
package, since they are still using the github.knative/pkg one.

Currently builds are suffixed with a timestamp and a string sort
operation determines the latest one. Long term we might want to limit
the history of a source aka what's the maximum size of this operation.
(Maybe something configurable in the space?) From what I can tell, this
pattern will continue when we move to tekton.

There are some errors being thrown when status setting. They seem to not
cause any failure. Further investigation is required.